pyproject.toml
[tool.poetry]
name = "encryption"
version = "0.1.0"
description = ""
authors = ["gitmanntoo <gitmanntoo@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
cryptography = "^44.0.0"
python-gnupg = "^0.5.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
